refactor(test): update unit and integration tests code
