+++ tests/src/Flyvemdm/Tests/CommonTestCase.php
@@ -30,8 +30,6 @@
  */
 namespace Flyvemdm\Tests;
 
-use Flyvemdm\Tests\TestingCommonTools;
-
 class CommonTestCase extends \Glpi\Tests\CommonTestCase {
 
    /**
@@ -122,7 +120,7 @@ class CommonTestCase extends \Glpi\Tests\CommonTestCase {
          $finalVersion = null;
       }
 
-      $finalInventory = (null !== $inventory) ? $inventory : TestingCommonTools::AgentXmlInventory($serial);
+      $finalInventory = (null !== $inventory) ? $inventory : CommonTestCase::AgentXmlInventory($serial);
 
       $input = [
          'entities_id'       => $_SESSION['glpiactive_entity'],
@@ -157,5 +155,299 @@ class CommonTestCase extends \Glpi\Tests\CommonTestCase {
       $this->boolean($fleet->isNewItem())->isFalse();
 
       return $fleet;
+   }
+
+   /**
+    * Create a file directly on DB and return the object and the filename generated
+    *
+    * @param $entityId
+    * @param null $userFilename
+    * @param int $version
+    * @return \PluginFlyvemdmFile
+    */
+   function createFlyvemdmDumbFile($entityId, $userFilename = null, $version = 1) {
+      global $DB;
+
+      // Create an file (directly in DB)
+      $fileName = ((null !== $userFilename) ? $userFilename : $this->getUniqueString());
+      $destination = $entityId . '/dummy_file_' . $fileName;
+      if (!is_dir($directory = FLYVEMDM_FILE_PATH . "/" . $entityId)) {
+         @mkdir($directory);
+      }
+      $fileSize = file_put_contents(FLYVEMDM_FILE_PATH . '/' . $destination, 'dummy');
+      $this->integer($fileSize)->isGreaterThan(0);
+      $fileTable = \PluginFlyvemdmFile::getTable();
+      $query = "INSERT INTO $fileTable (
+         `name`,
+         `source`,
+         `entities_id`,
+         `version`
+      ) VALUES (
+         '$fileName',
+         '$destination',
+         '$entityId',
+         '$version'
+      )";
+
+      try {
+         $DB->query($query);
+         $mysqlError = $DB->error();
+         $flyvemdmFile = new \PluginFlyvemdmFile();
+         $flyvemdmFile->getFromDBByCrit(['name' => $fileName]);
+         $this->boolean($flyvemdmFile->isNewItem())->isFalse($mysqlError);
+         return $flyvemdmFile;
+      } catch (\Exception $e){
+         echo $e->getMessage();
+         $this->stop();
+      }
+   }
+
+   /**
+    * Fake XML data for inventory
+    * @param $serial
+    * @param string $macAddress
+    * @param string $deviceId
+    * @param string $uuid
+    * @return string
+    */
+   public static function AgentXmlInventory($serial, $macAddress = '', $deviceId = '', $uuid = '') {
+      $uuid = ($uuid) ? $uuid : '1d24931052f35d92';
+      $macAddress = ($macAddress) ? $macAddress : '02:00:00:00:00:00';
+      $deviceId = ($deviceId) ? $deviceId : $serial . "_" . $macAddress;
+      $xml = "<?xml version='1.0' encoding='utf-8' standalone='yes'?>
+            <REQUEST>
+              <QUERY>INVENTORY</QUERY>
+              <VERSIONCLIENT>FlyveMDM-Agent_v1.0</VERSIONCLIENT>
+              <DEVICEID>" . $deviceId . "</DEVICEID>
+              <CONTENT>
+                <ACCESSLOG>
+                  <LOGDATE>" . date("Y-m-d H:i:s") . "</LOGDATE>
+                  <USERID>N/A</USERID>
+                </ACCESSLOG>
+                <ACCOUNTINFO>
+                  <KEYNAME>TAG</KEYNAME>
+                  <KEYVALUE/>
+                </ACCOUNTINFO>
+                <HARDWARE>
+                  <DATELASTLOGGEDUSER>09/11/17</DATELASTLOGGEDUSER>
+                  <LASTLOGGEDUSER>jenkins</LASTLOGGEDUSER>
+                  <NAME>" . $serial . "</NAME>
+                  <OSNAME>Android</OSNAME>
+                  <OSVERSION>6.0</OSVERSION>
+                  <ARCHNAME>aarch64</ARCHNAME>
+                  <UUID>" . $uuid . "</UUID>
+                  <MEMORY>1961</MEMORY>
+                </HARDWARE>
+                <BIOS>
+                  <BDATE>09/11/17</BDATE>
+                  <BMANUFACTURER>bq</BMANUFACTURER>
+                  <MMANUFACTURER>bq</MMANUFACTURER>
+                  <SMODEL>Aquaris M10 FHD</SMODEL>
+                  <SSN>" . $serial . "</SSN>
+                </BIOS>
+                <MEMORIES>
+                  <DESCRIPTION>Memory</DESCRIPTION>
+                  <CAPACITY>1961</CAPACITY>
+                </MEMORIES>
+                <INPUTS>
+                  <CAPTION>Touch Screen</CAPTION>
+                  <DESCRIPTION>Touch Screen</DESCRIPTION>
+                  <TYPE>FINGER</TYPE>
+                </INPUTS>
+                <SENSORS>
+                  <NAME>ACCELEROMETER</NAME>
+                  <NAME>MTK</NAME>
+                  <TYPE>ACCELEROMETER</TYPE>
+                  <POWER>0.13</POWER>
+                  <VERSION>3</VERSION>
+                </SENSORS>
+                <SENSORS>
+                  <NAME>LIGHT</NAME>
+                  <NAME>MTK</NAME>
+                  <TYPE>Unknow</TYPE>
+                  <POWER>0.13</POWER>
+                  <VERSION>1</VERSION>
+                </SENSORS>
+                <SENSORS>
+                  <NAME>ORIENTATION</NAME>
+                  <NAME>MTK</NAME>
+                  <TYPE>Unknow</TYPE>
+                  <POWER>0.25</POWER>
+                  <VERSION>3</VERSION>
+                </SENSORS>
+                <SENSORS>
+                  <NAME>MAGNETOMETER</NAME>
+                  <NAME>MTK</NAME>
+                  <TYPE>MAGNETIC FIELD</TYPE>
+                  <POWER>0.25</POWER>
+                  <VERSION>3</VERSION>
+                </SENSORS>
+                <DRIVES>
+                  <VOLUMN>/system</VOLUMN>
+                  <TOTAL>1487</TOTAL>
+                  <FREE>72</FREE>
+                </DRIVES>
+                <DRIVES>
+                  <VOLUMN>/storage/emulated/0</VOLUMN>
+                  <TOTAL>12529</TOTAL>
+                  <FREE>8322</FREE>
+                </DRIVES>
+                <DRIVES>
+                  <VOLUMN>/data</VOLUMN>
+                  <TOTAL>12529</TOTAL>
+                  <FREE>8322</FREE>
+                </DRIVES>
+                <DRIVES>
+                  <VOLUMN>/cache</VOLUMN>
+                  <TOTAL>410</TOTAL>
+                  <FREE>410</FREE>
+                </DRIVES>
+                <CPUS>
+                  <NAME>AArch64 Processor rev 3 (aarch64)</NAME>
+                  <SPEED>1500</SPEED>
+                </CPUS>
+                <SIMCARDS>
+                  <STATE>SIM_STATE_UNKNOWN</STATE>
+                </SIMCARDS>
+                <VIDEOS>
+                  <RESOLUTION>1920x1128</RESOLUTION>
+                </VIDEOS>
+                <CAMERAS>
+                  <RESOLUTIONS>3264x2448</RESOLUTIONS>
+                </CAMERAS>
+                <CAMERAS>
+                  <RESOLUTIONS>2880x1728</RESOLUTIONS>
+                </CAMERAS>
+                <NETWORKS>
+                  <TYPE>WIFI</TYPE>
+                  <MACADDR>" . $macAddress . "</MACADDR>
+                  <SPEED>65</SPEED>
+                  <BSSID>aa:5b:78:78:52:7e</BSSID>
+                  <SSID>aa:5b:78:78:52:7e</SSID>
+                  <IPGATEWAY>172.20.10.1</IPGATEWAY>
+                  <IPADDRESS>172.20.10.3</IPADDRESS>
+                  <IPMASK>0.0.0.0</IPMASK>
+                  <IPDHCP>172.20.10.1</IPDHCP>
+                </NETWORKS>
+                <ENVS>
+                  <KEY>SYSTEMSERVERCLASSPATH</KEY>
+                  <VAL>/system/framework/services.jar:/system/framework/ethernet-service.jar:/system/framework/wifi-service.jar</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ANDROID_SOCKET_zygote</KEY>
+                  <VAL>11</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ANDROID_DATA</KEY>
+                  <VAL>/data</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>PATH</KEY>
+                  <VAL>/sbin:/vendor/bin:/system/sbin:/system/bin:/system/xbin</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ANDROID_ASSETS</KEY>
+                  <VAL>/system/app</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ANDROID_ROOT</KEY>
+                  <VAL>/system</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ASEC_MOUNTPOINT</KEY>
+                  <VAL>/mnt/asec</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>LD_PRELOAD</KEY>
+                  <VAL>libdirect-coredump.so</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ANDROID_BOOTLOGO</KEY>
+                  <VAL>1</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>BOOTCLASSPATH</KEY>
+                  <VAL>/system/framework/core-libart.jar:/system/framework/conscrypt.jar:/system/framework/okhttp.jar:/system/framework/core-junit.jar:/system/framework/bouncycastle.jar:/system/framework/ext.jar:/system/framework/framework.jar:/system/framework/telephony-common.jar:/system/framework/voip-common.jar:/system/framework/ims-common.jar:/system/framework/apache-xml.jar:/system/framework/org.apache.http.legacy.boot.jar:/system/framework/mediatek-common.jar:/system/framework/mediatek-framework.jar:/system/framework/mediatek-telephony-common.jar:/system/framework/dolby_ds2.jar:/system/framework/dolby_ds1.jar</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ANDROID_PROPERTY_WORKSPACE</KEY>
+                  <VAL>9,0</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>EXTERNAL_STORAGE</KEY>
+                  <VAL>/sdcard</VAL>
+                </ENVS>
+                <ENVS>
+                  <KEY>ANDROID_STORAGE</KEY>
+                  <VAL>/storage</VAL>
+                </ENVS>
+                <JVMS>
+                  <NAME>Dalvik</NAME>
+                  <LANGUAGE>en_GB</LANGUAGE>
+                  <VENDOR>The Android Project</VENDOR>
+                  <RUNTIME>0.9</RUNTIME>
+                  <HOME>/system</HOME>
+                  <VERSION>2.1.0</VERSION>
+                  <CLASSPATH>.</CLASSPATH>
+                </JVMS>
+                <BATTERIES>
+                  <CHEMISTRY>Li-ion</CHEMISTRY>
+                  <TEMPERATURE>23.0c</TEMPERATURE>
+                  <VOLTAGE>3.745V</VOLTAGE>
+                  <LEVEL>60%</LEVEL>
+                  <HEALTH>Good</HEALTH>
+                  <STATUS>Not charging</STATUS>
+                </BATTERIES>
+              </CONTENT>
+            </REQUEST>";
+      return base64_encode($xml);
+   }
+
+   /**
+    * Create an application (directly in DB) because we are not uploading any file
+    * @return \PluginFlyvemdmPackage
+    */
+   protected function createFlyvemdmDumbPackage($entityId, $filename = null, $version = '1.0.5') {
+      global $DB;
+
+      // Create an file (directly in DB)
+      $uniqueString = ((null !== $filename) ? $filename : $this->getUniqueString());
+      $dumbPackageName = 'com.domain.' . $uniqueString . '.application';
+      $destination = 'flyvemdm/package/' .$entityId . '/123456789_application_' . $uniqueString . '.apk';
+      if (!is_dir($directory = FLYVEMDM_PACKAGE_PATH . "/" . $entityId)) {
+         @mkdir($directory);
+      }
+      $fileSize = file_put_contents(GLPI_PLUGIN_DOC_DIR . '/' . $destination, 'dummy');
+      $this->integer($fileSize)->isGreaterThan(0);
+      $packageTable = \PluginFlyvemdmPackage::getTable();
+      $query = "INSERT INTO $packageTable (
+         `package_name`,
+         `alias`,
+         `version`,
+         `filename`,
+         `entities_id`,
+         `dl_filename`,
+         `icon`
+      ) VALUES (
+         '$dumbPackageName',
+         'application',
+         '$version',
+         '$destination',
+         '$entityId',
+         'application_" . $uniqueString . ".apk',
+         ''
+      )";
+
+      try {
+         $DB->query($query);
+         $mysqlError = $DB->error();
+         $flyvemdmPackage = new \PluginFlyvemdmPackage();
+         $flyvemdmPackage->getFromDBByCrit(['package_name' => $dumbPackageName]);
+         $this->boolean($flyvemdmPackage->isNewItem())->isFalse($mysqlError);
+         return $flyvemdmPackage;
+      } catch (\Exception $e){
+         echo $e->getMessage();
+         $this->stop();
+      }
    }
 }
