+++ tests/suite-integration/PluginFlyvemdmFleet.php
@@ -92,7 +92,7 @@ class PluginFlyvemdmFleet extends CommonTestCase {
             'lastname'          => 'Doe',
             'version'           => \PluginFlyvemdmAgent::MINIMUM_ANDROID_VERSION . '.0',
             'type'              => 'android',
-            'inventory'         => TestingCommonTools::AgentXmlInventory($serial),
+            'inventory'         => CommonTestCase::AgentXmlInventory($serial),
          ]
       );
       $this->boolean($agent->isNewItem())
