+++ tests/suite-unit/PluginFlyvemdmPolicyBoolean.php
@@ -31,7 +31,7 @@
 
 namespace tests\units;
 
-use Glpi\Tests\CommonTestCase;
+use Flyvemdm\Tests\CommonTestCase;
 
 class PluginFlyvemdmPolicyBoolean extends CommonTestCase {
 
@@ -135,8 +135,8 @@ class PluginFlyvemdmPolicyBoolean extends CommonTestCase {
 
    /**
     * @dataProvider filterStatusProvider
-    * @param unknown $status
-    * @param unknown $expected
+    * @param mixed $status
+    * @param mixed $expected
     */
    public function testFilterStatus($status, $expected) {
       $policy = new \PluginFlyvemdmPolicy();
