+++ tests/suite-unit/PluginFlyvemdmPolicyDeployapplication.php
@@ -31,7 +31,7 @@
 
 namespace tests\units;
 
-use Glpi\Tests\CommonTestCase;
+use Flyvemdm\Tests\CommonTestCase;
 
 class PluginFlyvemdmPolicyDeployapplication extends CommonTestCase {
 
@@ -41,9 +41,6 @@ class PluginFlyvemdmPolicyDeployapplication extends CommonTestCase {
       'type_data' => '',
       'unicity'   => 0,
    ];
-
-   private $packageName = 'com.domain.author.application.apk';
-   private $filename = 'application.apk';
 
    /**
     * @return array
@@ -97,7 +94,7 @@ class PluginFlyvemdmPolicyDeployapplication extends CommonTestCase {
    public function testCreatePolicy($data, $expected) {
       list($policy) = $this->createNewPolicyInstance();
       if ($data[2] === true) {
-         $item = $this->createAppInDB();
+         $item = $this->createFlyvemdmDumbPackage(0);
          $data[2] = $item->getID();
       }
       $success = $policy->integrityCheck($data[0], $data[1], $data[2]);
@@ -109,49 +106,13 @@ class PluginFlyvemdmPolicyDeployapplication extends CommonTestCase {
    }
 
    /**
-    * Create an application (directly in DB) because we are not uploading any file
-    * @return \PluginFlyvemdmPackage
-    */
-   private function createAppInDB() {
-      global $DB;
-
-      $uniqid = uniqid();
-      $table_file = \PluginFlyvemdmPackage::getTable();
-      $query = "INSERT INTO `$table_file` (
-        `package_name`,
-        `alias`,
-        `version`,
-        `filename`,
-        `entities_id`,
-        `dl_filename`,
-        `icon`
-        ) VALUES (
-        '" . $uniqid . $this->packageName . "',
-        'application',
-        '1.0.5',
-        '0/" . $uniqid . $this->filename . "',
-        '0',
-        '" . $this->filename . "',
-        ''
-        )";
-      $result = $DB->query($query);
-      $this->boolean($result)->isTrue();
-
-      $file = new \PluginFlyvemdmPackage();
-      $file->getFromDB($DB->insert_id());
-      $this->boolean($file->isNewItem())->isFalse();
-
-      return $file;
-   }
-
-   /**
     * @tags testGetMqttMessage
     */
    public function testGetMqttMessage() {
       list($policy) = $this->createNewPolicyInstance();
 
       $this->boolean($policy->getMqttMessage(null, null, null))->isFalse();
-      $item = $this->createAppInDB();
+      $item = $this->createFlyvemdmDumbPackage(0);
       $value = '{"remove_on_delete":0}';
       $result = $policy->getMqttMessage($value, $item->getType(), $item->getID());
       $this->array($result)->hasKeys(['id', 'versionCode', $this->dataField['symbol']])
@@ -167,7 +128,7 @@ class PluginFlyvemdmPolicyDeployapplication extends CommonTestCase {
       list($policy) = $this->createNewPolicyInstance();
       $mockInstance = $this->newMockInstance('\PluginFlyvemdmFleet');
       $mockInstance->getMockController()->getID = 1;
-      $application = $this->createAppInDB();
+      $application = $this->createFlyvemdmDumbPackage(0);
       $this->boolean($policy->unicityCheck(null, \PluginFlyvemdmPackage::class, $application->getID(),
          $mockInstance))->isTrue();
       // TODO: finish this test
@@ -185,7 +146,7 @@ class PluginFlyvemdmPolicyDeployapplication extends CommonTestCase {
       $this->string($_SESSION["MESSAGE_AFTER_REDIRECT"][1][0])
          ->isEqualTo('The application does not exists');
       unset($_SESSION["MESSAGE_AFTER_REDIRECT"]); // to clear the buffer
-      $application = $this->createAppInDB();
+      $application = $this->createFlyvemdmDumbPackage(0);
       $this->boolean($policy->conflictCheck(null, $packageClass, $application->getID(),
          $mockInstance))->isTrue();
       // TODO: finish this test
@@ -198,7 +159,7 @@ class PluginFlyvemdmPolicyDeployapplication extends CommonTestCase {
       list($policy) = $this->createNewPolicyInstance();
       $mockInstance = $this->newMockInstance(\PluginFlyvemdmFleet::class);
       $mockInstance->getMockController()->getID = 1;
-      $appInDB = $this->createAppInDB();
+      $appInDB = $this->createFlyvemdmDumbPackage(0);
 
       // check integrity
       $this->boolean($policy->pre_unapply(null, null, null, $mockInstance))->isFalse();
