+++ tests/suite-unit/PluginFlyvemdmPolicyString.php
@@ -31,7 +31,7 @@
 
 namespace tests\units;
 
-use Glpi\Tests\CommonTestCase;
+use Flyvemdm\Tests\CommonTestCase;
 use PluginFlyvemdmPolicy;
 
 class PluginFlyvemdmPolicyString extends CommonTestCase {
@@ -98,8 +98,8 @@ class PluginFlyvemdmPolicyString extends CommonTestCase {
 
    /**
     * @dataProvider filterStatusProvider
-    * @param unknown $status
-    * @param unknown $expected
+    * @param mixed $status
+    * @param mixed $expected
     */
    public function testFilterStatus($status, $expected) {
       $policy = new \PluginFlyvemdmPolicy();
